Update window creation (#2028)
* add ui/window

* move window and optimize code

* cleanup
